+++ examples/complete-vpc/outputs.tf
@@ -67,20 +67,3 @@ output "this_customer_gateway" {
   description = "Map of Customer Gateway attributes"
   value       = module.vpc.this_customer_gateway
 }
-
-//
-//# VPC endpoints
-//output "vpc_endpoint_ec2_id" {
-//  description = "The ID of VPC endpoint for EC2"
-//  value       = "${module.vpc.vpc_endpoint_ec2_id}"
-//}
-//
-//output "vpc_endpoint_ec2_network_interface_ids" {
-//  description = "One or more network interfaces for the VPC Endpoint for EC2."
-//  value = ["${module.vpc.vpc_endpoint_ec2_network_interface_ids}"]
-//}
-//
-//output "vpc_endpoint_ec2_dns_entry" {
-//  description = "The DNS entries for the VPC Endpoint for EC2."
-//  value = ["${module.vpc.vpc_endpoint_ec2_dns_entry}"]
-//}